Generate compiled regular expressions rather than compile at runtime
The use of uncompiled regexes creates enormous amounts of objects and
CPU use. All patterns are static - generate them up front as static
variables in the compiler. Remove the PatternMatches() helper as it is
no longer used.

Add special casing for common patterns to avoid the use of regex for
trivial actions, such as prefix matches. Hits the common OpenAPI v2 and
v3 restrictions.