Only add requestBody when the content type is json
We don't support xml or form data encoding, so no need to add it as a
parameter in the client.

I don't know if we should have smarter matching for other media types
(text/json).
I don't know if we should be setting the Accept header in the server.

This was causing duplicate parameters and breaking the build.