Reorganize tests.
The top-level tests now use a "lib" version of Gnostic
to run in a single process, which allows them to run in
test environments that require that.

Plugin and extension tests are moved into their respective
directories and are run from "make test" which is also
used by .travis.yaml to run these tests in CI. One sample
app (petstore-builder) is also tested in this way.

To avoid confusing the "make" tool, the "test" directory
is renamed "testdata".