updates and improvements to schema-based parsing.
The biggest change here is that the models for JSON references
are now "open" and allow other keys to be specified alongside
"$ref". Following the spec, these are ignored, where previously
they triggered error messages that caused some API descriptions
to fail to parse.

This commit also contains small tweaks to the openapi-3.0.json
schema that improves model coverage and a small refinement to
error printing (usage information is only printed for usage-
related errors, fixing a problem introduced in pr#161).