Fix a bug in type generation.
Schemas that are nested in other schemas are discovered
as those schemas are expanded. This wasn't an issue
for the OpenAPI schemas, but the Discovery schema has a
"mediaUpload" schema that contains a "protocols" property
with an embedded schema which in turn contains two more
schemas (for the "simple" and "resumable" properties).
These schemas were being discovered when the "protocols"
schema was being expanded and were being added to a map
while the expansion was iterating through the map. This
caused instability - sometimes the schema types were generated,
and sometimes they weren't.