Put output files in the same directory as their associated input files. (#180)
* Put output files in the same directory as their associated input files.

When a directory name is specified as a gnostic output directory,
files were previously written directly to that directory. With this
change, files will be written into the same directory as the source
file that gnostic read. This is most helpful when gnostic is run
over a collection of files using a tool like "find":

```find -name "openapi.yaml" --pb_out=.```

Previously this would write all of the outputs onto a single file
(overwriting all but the last file written). Now this writes the
outputs in the same directory as the source files, e.g. in
`myapi/2.0/openapi.pb` (for an API spec at `myapi/2.0/openapi.yaml`).

This required updates to tests to match the new behavior, and the
swagger.text file that is deleted probably should have never been
checked in.

* Create intermediate directories if needed.