Generate jsonschema/base.go (a compiled-in version of jsonschema/schema.json).
This commit uses "go generate" to embed jsonschema/schema.json in the
jsonschema package so that it is no longer read at runtime. Prior to this
reads would fail if gnostic was not installed in the usual place in a
user's GOPATH, and as Go module usage has become common, this has led
to common build failures.

Here we also replace the "COMPILE-PROTOS.sh" with "go generate ./...", which
causes both COMPILE-PROTOS.sh to be run (since it is called for in gnostic.go)
and jsonschema/generate-base.go to be run (due to the go:generate comment in
jsonschema/reader.go).