Use tags for grouping + add title and description as config params
A new tag is created per services and added to each of it's operations.

The title and description has been added as standard params and can be set as key value params with either --openapi_out or --openapi_opt.
E.g.:
--openapi_opt=title=API,description="testing testing",version=1.2.3
--openapi_out=json=true:.

Title and description of the document is handled a bit differently:
 - A service's name now becomes it's tag
 - Comments will be used for description instead of summary
 - If doc title and description is set, it is never overwritten
 - If only one service is being parsed:
    - If doc has no title, use the service's tag name + " API"
    - If doc has no description, use service tag description
    - Remove the tag description