Fix query parameters
This fixes issues described in #275.
Specifically:
 - types (everything was set to string)
 - message and sub message handling

Changes was done according to the note in annotations.proto:

Note that fields which are mapped to URL query parameters must have a primitive type or a repeated primitive type or a non-repeated message type.
In the case of a repeated type, the parameter can be repeated in the URL as ...?param=A&param=B.
In the case of a message type, each field of the message is mapped to a separate parameter, such as ...?foo.a=A&foo.b=B&foo.c=C.

The ability to handle sub sub messages was tested with Envoy first.