Migrate gnostic to use gnostic-models (#400)
In order to avoid a panic at runtime in a downstream project that pulls in gnostic and gnostic-models, create a new `<proto>.pbalias.go` file which uses type aliases and variables to point to the equivalent protobuf types in gnostic-models. This will prevent any API breakage for code using the previous types from gnostic, and allow for this logic to be maintained in one place going forward.